[Feature/57] session service 개발 중

See merge request s09-bigdata-recom-sub2/S09P22A402!17

Former-commit-id: 47fada4cebc56a1e4d6d35d796f02733b1aaa82e

+++ Backend/src/main/java/com/ontheblock/www/session/dto/SessionRequest.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class SessionRequest {
     Long memberId;
 
-    Integer instrumentId;
+    Long instrumentId;
 }
